Removed interface IndicatorCriteriaDAO
